+++ src/main/java/com/codeshop/persephone/propositions/PropositionRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PropositionRepository extends JpaRepository<ProposedGroup, Long> {
+public interface PropositionRepository extends JpaRepository<ProposedGroup, Integer> {
 }
